테스트 - reducer 테스트 추가
store 파일에서 reducer 파일로 분리

updateTaskTitle 액션 테스트
- change a new task title

addTask 액션 테스트
- with task title
  - appends a new task into tasks
  - clears task title
- without task title
  - doesn't work
